refactor: 코드 수정
[Enum 클래스 위치 변경, Entity -> Entity/type,
createDate-> createDateTime 으로 변경
createDateTime 컬럼에 updatable 추가]
