+++ user-api/src/main/java/org/ecommerce/userapi/entity/Users.java
@@ -2,6 +2,8 @@ package org.ecommerce.userapi.entity;
 
 import java.time.LocalDateTime;
 
+import org.ecommerce.userapi.entity.type.Gender;
+import org.ecommerce.userapi.entity.type.Status;
 import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
@@ -47,16 +49,16 @@ public class Users {
 	@Column(name = "phone_number", nullable = false)
 	private String phoneNumber;
 
-	@Column(name = "is_deleted")
-	private boolean isDeleted;
-
 	@CreationTimestamp
-	@Column(name = "create_date")
-	private LocalDateTime createDate;
+	@Column(name = "create_datetime",updatable = false)
+	private LocalDateTime createDatetime;
+
+	@Column(name = "is_deleted")
+	private Boolean isDeleted;
 
 	@UpdateTimestamp
-	@Column(name = "update_date")
-	private LocalDateTime updateDate;
+	@Column(name = "update_datetime")
+	private LocalDateTime updateDatetime;
 
 	@ColumnDefault("0")
 	@Column(name = "beanpay")
