+++ user-api/src/main/java/org/ecommerce/userapi/entity/type/Gender.java
@@ -1,4 +1,4 @@
-package org.ecommerce.userapi.entity;
+package org.ecommerce.userapi.entity.type;
 
 import lombok.Getter;
